Adding Tour Points and fixed Normals on GLTFs
added a component to selectively flip normals on GLTF files to fix the
upside down normals on normal maps saved for direct X format.

added tour point place holders and two forms of text for display.  Html
shader text and normal font map text.  The HTML works better from a
rendering perspective.  It may not be as performant as the standard
aframe font.